feat: implement utility functions for parameter validation and error handling
