+++ promotions_oneclick_mall/views.py
@@ -1,16 +1,19 @@
 import os
 import secrets
-from decimal import Decimal, InvalidOperation
 
 from django.shortcuts import render
 from django.views.decorators.http import require_GET
-from transbank.error.transbank_error import TransbankError
+from core.utils import (
+    render_error,
+    required_decimal_param,
+    required_integer_param,
+    required_text_param,
+)
 from transbank.webpay.oneclick.mall_bin_info import MallBinInfo
 from transbank.webpay.oneclick.mall_inscription import MallInscription
 from transbank.webpay.oneclick.mall_transaction import MallTransaction
 from transbank.webpay.oneclick.request import MallTransactionAuthorizeDetails
 
-ERROR_TEMPLATE = "error_pages/general_error.html"
 APP_PATH = "/promotions-oneclick-mall"
 APP_NAME = "promotions_oneclick_mall"
 APP_TITLE = "Webpay Oneclick Mall Promociones"
@@ -46,44 +49,6 @@ def get_transbank_transaction():
 def get_transbank_bin_info():
     commerce_code, api_key = get_commerce_options()
     return MallBinInfo.build_for_integration(commerce_code, api_key)
-
-
-def required_text_param(request, key):
-    value = request.GET.get(key)
-    if value:
-        return value
-    raise ValueError(f"El parámetro {key} es obligatorio.")
-
-
-def required_decimal_param(request, key):
-    value = required_text_param(request, key)
-    try:
-        return Decimal(value)
-    except InvalidOperation as exc:
-        raise ValueError(f"El parámetro {key} debe ser numérico.") from exc
-
-
-def required_integer_param(request, key):
-    value = required_text_param(request, key)
-    try:
-        return int(value)
-    except ValueError as exc:
-        raise ValueError(f"El parámetro {key} debe ser un entero.") from exc
-
-
-def render_error(request, exception):
-    return render(request, ERROR_TEMPLATE, {"error": displayable_error_message(exception)})
-
-
-def displayable_error_message(exception):
-    current = exception
-
-    while current:
-        if isinstance(current, TransbankError):
-            return str(current)
-        current = current.__cause__ or current.__context__
-
-    return "Ocurrió un error inesperado al procesar la operación."
 
 
 @require_GET
